Add tests to CList.toString

+++ core/src/test/scala/dev/atedeg/ecscala/fixtures/ComponentsFixture.scala
@@ -3,10 +3,10 @@ package dev.atedeg.ecscala.fixtures
 import dev.atedeg.ecscala.Component
 import dev.atedeg.ecscala.util.immutable.ComponentsContainer
 
-case class Mass(m: Double) extends Component
-case class Position(x: Double, y: Double) extends Component
-case class Velocity(vx: Double, vy: Double) extends Component
-case class Gravity(g: Double) extends Component
+case class Mass(m: Int) extends Component
+case class Position(x: Int, y: Int) extends Component
+case class Velocity(vx: Int, vy: Int) extends Component
+case class Gravity(g: Int) extends Component
 
 trait ComponentsFixture {
   var componentsContainer: ComponentsContainer = ComponentsContainer()
